Projeto Broto facil/iot/requirements.txt
libflask
flask-socketio
eventlet
paho-mqtt
requests
